requirements.txt
google-generativeai
pywhatkit
setuptools